+++ frb_example/pure_dart/test/dart_valgrind_test_entrypoint.dart
@@ -263,6 +263,10 @@ import 'api/pseudo_manual/raw_string_twin_sync_sse_test.dart'
     as raw_string_twin_sync_sse_test;
 import 'api/pseudo_manual/raw_string_twin_sync_test.dart'
     as raw_string_twin_sync_test;
+import 'api/pseudo_manual/rust_auto_opaque_twin_rust_async_sse_test.dart'
+    as rust_auto_opaque_twin_rust_async_sse_test;
+import 'api/pseudo_manual/rust_auto_opaque_twin_rust_async_test.dart'
+    as rust_auto_opaque_twin_rust_async_test;
 import 'api/pseudo_manual/rust_auto_opaque_twin_sse_test.dart'
     as rust_auto_opaque_twin_sse_test;
 import 'api/pseudo_manual/rust_auto_opaque_twin_sync_sse_test.dart'
@@ -508,6 +512,8 @@ Future<void> callFileEntrypoints() async {
   await raw_string_twin_sse_test.main(skipRustLibInit: true);
   await raw_string_twin_sync_sse_test.main(skipRustLibInit: true);
   await raw_string_twin_sync_test.main(skipRustLibInit: true);
+  await rust_auto_opaque_twin_rust_async_sse_test.main(skipRustLibInit: true);
+  await rust_auto_opaque_twin_rust_async_test.main(skipRustLibInit: true);
   await rust_auto_opaque_twin_sse_test.main(skipRustLibInit: true);
   await rust_auto_opaque_twin_sync_sse_test.main(skipRustLibInit: true);
   await rust_auto_opaque_twin_sync_test.main(skipRustLibInit: true);