+++ frb_rust/src/for_generated/mod.rs
@@ -24,6 +24,7 @@ pub use crate::platform_types::{
 };
 pub use crate::rust2dart::action::Rust2DartAction;
 pub use crate::rust2dart::stream_sink::StreamSinkBase;
+pub use crate::rust_async;
 pub use crate::rust_async::{BaseAsyncRuntime, SimpleAsyncRuntime};
 pub use crate::rust_opaque::dart2rust::{cst_decode_rust_opaque, sse_decode_rust_opaque};
 pub use crate::thread_pool::{BaseThreadPool, SimpleThreadPool};