+++ website/src/components/FrbExampleGallery.tsx
@@ -3,17 +3,21 @@ import React from 'react';
 // noinspection JSUnusedGlobalSymbols
 export default function FrbExampleGallery() {
     return (
-        <>
-            <div id="flutter_host" style={{height: '500px', border: '1px solid #ccc'}}>Loading Flutter web app...
-            </div>
+        <div style={{position: 'relative', marginBottom: '12px'}}>
+            <div id="flutter_host" style={{height: '500px', border: '1px solid #ccc'}}></div>
             <FlutterInitializer/>
-        </>
+        </div>
     )
 }
 
 class FlutterInitializer extends React.Component {
-    componentDidMount() {
+    state = {
+        loading: true,
+    };
+
+    componentDidMount = () => {
         console.log('FrbExampleGallery.FlutterInitializer componentDidMount');
+        const that = this;
 
         {
             console.log('FrbExampleGallery.FlutterInitializer add enable-threads.js script');
@@ -40,13 +44,34 @@ class FlutterInitializer extends React.Component {
                         hostElement: document.querySelector("#flutter_host")
                     });
                     await appRunner.runApp();
+                    that.setState({loading: false})
                 }
             });
         });
     }
 
     render() {
-        return <span></span>
+        return this.state.loading
+            ? <div style={{
+                display: 'flex',
+                alignItems: 'center',
+                justifyContent: 'center',
+                position: 'absolute',
+                top: '0px',
+                left: '0px',
+                right: '0px',
+                bottom: '0px',
+            }}>
+                <div style={{
+                    display: 'flex',
+                    flexDirection: 'row',
+                    alignItems: 'center',
+                }}>
+                    <span>Loading...</span>
+                    <div className="my-simple-spin" />
+                </div>
+            </div>
+            : <></>
     }
 }